+++ index.d.ts
@@ -24,7 +24,7 @@ export {ConditionalPick} from './source/conditional-pick';
 export {UnionToIntersection} from './source/union-to-intersection';
 export {Stringified} from './source/stringified';
 export {FixedLengthArray} from './source/fixed-length-array';
-export {IterableElement} from './source/iterable';
+export {IterableElement} from './source/iterable-element';
 
 // Miscellaneous
 export {PackageJson} from './source/package-json';
